Add files via upload
These are the documents for the Milwaukee Brewers Blogging website I created in my Web Application Development class. Will need to update the login information for the database in the mysqli_connect.php file.
Contact me for admin login. I have made the website so users cannot see comments until logged in and the administrator is the only one that can add new posts.